fixed order of migration and removed stuff on integers

+++ backend/db/migrations/20240821194722-create-entry.js
@@ -20,11 +20,11 @@ module.exports = {
         defaultValue: Sequelize.literal('CURRENT_TIMESTAMP')
       },
       overallMood: {
-        type: Sequelize.INTEGER(10),
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       iconId: {
-        type: Sequelize.INTEGER(5),
+        type: Sequelize.INTEGER,
         allowNull: false,
         references: {
           model: 'Icons'
