+++ backend/db/migrations/20240822015916-create-entry-level.js
@@ -30,7 +30,7 @@ module.exports = {
         }
       },
       rating: {
-        type: Sequelize.INTEGER(10),
+        type: Sequelize.INTEGER,
         allowNull:false
       },
       createdAt: {
